Some Home page editing and added 3 pages
The home page was edited by adding better facebook integration. I have
made three pages for the About Us section of the website. Including;
History, Special Guests and Life members (more editing for these pages
needs to be done). Images and context was uploaded and the navigation
bar was implemented.