Site Consistency and adjustments to About Us pages
Changes in the main header font and sizes, made it consistent throughout
all pages to make the website look neater. also did the same for the
headings on the navigation side bar and made the content text abit
larger so it is easier to read.

Made adjustments to the about us pages for overall look.